Move DataConstants and create IdentityConstants
