+++ PawMates/PawMates.Core/Models/PetViewModels/PetFormViewModel.cs
@@ -2,7 +2,7 @@
 using PawMates.Infrastructure.Data.Enums;
 using PawMates.Infrastructure.Data.Models;
 using System.ComponentModel.DataAnnotations;
-using static PawMates.Infrastructure.Data.DataConstants;
+using static PawMates.Infrastructure.Data.DataConstants.DataConstants;
 
 namespace PawMates.Core.Models.PetViewModels
 {
