+++ PawMates/PawMates.Core/Services/PetService/PetService.cs
@@ -3,7 +3,6 @@ using PawMates.Core.Contracts.PetInterface;
 using PawMates.Core.Models.EventViewModels;
 using PawMates.Core.Models.PetViewModels;
 using PawMates.Core.Services.EventService;
-using PawMates.Infrastructure.Data;
 using PawMates.Infrastructure.Data.Common;
 using PawMates.Infrastructure.Data.Models;
 using PawMates.Core.Models.PetViewModels;
@@ -12,8 +11,9 @@ using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using static PawMates.Infrastructure.Data.DataConstants;
+using static PawMates.Infrastructure.Data.DataConstants.DataConstants;
 using System;
+using PawMates.Infrastructure.Data.DataConstants;
 
 namespace PawMates.Core.Services.PetService
 {
