+++ PawMates/PawMates.Infrastructure/Data/Models/PetType.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using System.ComponentModel.DataAnnotations;
-using static PawMates.Infrastructure.Data.DataConstants;
+using static PawMates.Infrastructure.Data.DataConstants.DataConstants;
 namespace PawMates.Infrastructure.Data.Models
 {
     public class PetType
