+++ PawMates/PawMates.Infrastructure/Data/Models/Post.cs
@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore;
 using PawMates.Infrastructure.Data.Enums;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using static PawMates.Infrastructure.Data.DataConstants;
+using static PawMates.Infrastructure.Data.DataConstants.DataConstants;
 
 namespace PawMates.Infrastructure.Data.Models
 {
